Split config-host into remove, list, add commands
This PR also adds convenient aliases for

 - remove as `rm`
 - list as `ls`
 - add as `a`

Also brings much needed cleanup.
